ZCS-8834:ABQ - rename SuspendDeviceRequest
