+++ soap/src/java/com/zimbra/soap/JaxbUtil.java
@@ -726,8 +726,8 @@ public final class JaxbUtil {
             com.zimbra.soap.admin.message.SetPasswordResponse.class,
             com.zimbra.soap.admin.message.SetServerOfflineRequest.class,
             com.zimbra.soap.admin.message.SetServerOfflineResponse.class,
-            com.zimbra.soap.admin.message.SuspendDeviceRequest.class,
-            com.zimbra.soap.admin.message.SuspendDeviceResponse.class,
+            com.zimbra.soap.admin.message.QuarantineDeviceRequest.class,
+            com.zimbra.soap.admin.message.QuarantineDeviceResponse.class,
             com.zimbra.soap.admin.message.SyncGalAccountRequest.class,
             com.zimbra.soap.admin.message.SyncGalAccountResponse.class,
             com.zimbra.soap.admin.message.UndeployZimletRequest.class,
@@ -1094,8 +1094,8 @@ public final class JaxbUtil {
             com.zimbra.soap.sync.message.RemoteWipeResponse.class,
             com.zimbra.soap.sync.message.RemoveDeviceRequest.class,
             com.zimbra.soap.sync.message.RemoveDeviceResponse.class,
-            com.zimbra.soap.sync.message.SuspendDeviceRequest.class,
-            com.zimbra.soap.sync.message.SuspendDeviceResponse.class,
+            com.zimbra.soap.sync.message.QuarantineDeviceRequest.class,
+            com.zimbra.soap.sync.message.QuarantineDeviceResponse.class,
             com.zimbra.soap.voice.message.ChangeUCPasswordRequest.class,
             com.zimbra.soap.voice.message.ChangeUCPasswordResponse.class,
             com.zimbra.soap.voice.message.GetUCInfoRequest.class,
